requirements.txt
django_ratelimit
django_chunked_upload
Django==3.2.5
sorl-thumbnail
pillow
django-filter
djangorestframework
djangorestframework-simplejwt
coreapi
django-user-visit
django-notifications-hq
django-top-notify